requirements.txt
# Laria Security Scanner - Python Dependencies

# Core dependencies
pyyaml>=6.0.1
click>=8.1.7
jinja2>=3.1.2
requests>=2.31.0

# Git operations
gitpython>=3.1.40

# Report generation
markdown>=3.5.1
beautifulsoup4>=4.12.2

# JSON/YAML processing
jsonschema>=4.20.0

# Logging and utilities
colorama>=0.4.6
tabulate>=0.9.0
python-dateutil>=2.8.2

# Testing (optional, for development)
pytest>=7.4.3
pytest-cov>=4.1.0

# SAST
semgrep>=1.59.0

# IaC
checkov>=3.0.0
